flight-inventory/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
pika
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
opentelemetry-instrumentation-sqlalchemy
